Break up image processing module, make everything into layers
We are refactoring our repo to make the model parallelizable so we can
train across multiple GPUs. This requires two things:

- Converting constant tensors into Parameter instances.
- Have those Parameter instances be properties of nn.Module subclasses.

When these two requirements are met PyTorch's DataParallel class knows
to distribute these tensors to all GPUs during training. If they are not
distributed, we get errors for tensors in different GPUs being combined.

The key things we needed to convert to layers were:

- Color transform methods between rgb and lab/hsv color spaces.
- msssim and psnr calculations for loss and evaluation functions.

The former, in layer form, are in the new `colors.py` module. The latter
are in the `image_processing.py` module. image_processing.py was
bloated, so we split remaining methods out:

- Methods for transposing images are in `transpose.py`.
- Methods for applying curves for CURL are in `curves.py`.

Additionally we moved methods for loading images right into the
DataLoader class as static methods.

This way each submodule is more focused and its intent is much clearer,
while accomplishing our goal of refactoring to layers.
